src/main/java/LeetCode/LongestPalindromicSubsequence_516/Solution.java
package LeetCode.LongestPalindromicSubsequence_516;

class Solution {

    public int longestPalindromeSubseq(String s) {
        int len = s.length();
        int[][] dp = new int[len][len];
        dp[len - 1][0] = 1;

        for (int i = len - 2; i >= 0; i--) {
            for (int j = i; j < len; j++) {
                if (s.charAt(i) == s.charAt(j)
                        && (j - i <= 2
                                || dp[i + 1][j - 1] > 0)) {

                    if (j - i <= 2) {
                        dp[i][j] = j - i + 1;
                    } else {
                        dp[i][j] = dp[i + 1][j - 1] + 2;

                    }
                } else {
                    dp[i][j] = Math.max(dp[i][j - 1], dp[i + 1][j]);
                }
            }
        }
        return dp[0][len - 1];
    }
}